refactor: add container spec manager
Add container spec manager.
