+++ daemon/mgr/container.go
@@ -3,17 +3,15 @@ package mgr
 import (
 	"context"
 	"fmt"
-	"strconv"
-	"strings"
 	"time"
 
 	"github.com/alibaba/pouch/apis/types"
 	"github.com/alibaba/pouch/ctrd"
 	"github.com/alibaba/pouch/daemon/meta"
+	"github.com/alibaba/pouch/daemon/spec"
 	"github.com/alibaba/pouch/pkg/collect"
 	"github.com/alibaba/pouch/pkg/randomid"
 
-	specs "github.com/opencontainers/runtime-spec/specs-go"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 )
@@ -123,18 +121,18 @@ func (cm *ContainerManager) Start(ctx context.Context, startCfg types.ContainerS
 	c.DetachKeys = startCfg.DetachKeys
 
 	// new a default spec.
-	spec, err := ctrd.NewDefaultSpec(ctx)
+	s, err := ctrd.NewDefaultSpec(ctx)
 	if err != nil {
 		return errors.Wrapf(err, "failed to generate spec: %s", c.ID)
 	}
 
-	fs := []setupFunc{setupProcess, setupNetwork, setupCap, setupNs, setupMounts}
-	for _, f := range fs {
-		if err = f.run(ctx, c, spec); err != nil {
+	for _, f := range spec.SetupFuncs() {
+		if err = f(ctx, c, s); err != nil {
 			return err
 		}
 	}
-	err = cm.Client.CreateContainer(ctx, c, spec)
+
+	err = cm.Client.CreateContainer(ctx, c, s)
 	if err == nil {
 		c.Status = types.RUNNING
 		c.StartedAt = time.Now()
@@ -218,82 +216,4 @@ func (cm *ContainerManager) containerInfo(s string) (*types.ContainerInfo, error
 	}
 
 	return ci, nil
-}
-
-type setupFunc func(ctx context.Context, c *types.ContainerInfo, s *specs.Spec) error
-
-func (sf setupFunc) run(ctx context.Context, c *types.ContainerInfo, s *specs.Spec) error {
-	return sf(ctx, c, s)
-}
-
-func setupNetwork(ctx context.Context, c *types.ContainerInfo, s *specs.Spec) error {
-	s.Hostname = c.Config.Hostname
-	//TODO setup network parameters
-	return nil
-}
-
-func setupCap(ctx context.Context, c *types.ContainerInfo, s *specs.Spec) error {
-	//TODO setup capabilities
-	return nil
-}
-
-func setupNs(ctx context.Context, c *types.ContainerInfo, s *specs.Spec) error {
-	//TODO setup ns
-	return nil
-}
-
-func setupMounts(ctx context.Context, c *types.ContainerInfo, s *specs.Spec) error {
-	mounts := s.Mounts
-	if c.Config.HostConfig == nil {
-		return nil
-	}
-	for _, v := range c.Config.HostConfig.Binds {
-		sd := strings.SplitN(v, ":", 2)
-		if len(sd) != 2 {
-			continue
-		}
-		mounts = append(mounts, specs.Mount{
-			Destination: sd[0],
-			Source:      sd[1],
-			Type:        "bind",
-			Options:     []string{"rbind"},
-		})
-	}
-	s.Mounts = mounts
-	return nil
-}
-
-func setupProcess(ctx context.Context, c *types.ContainerInfo, s *specs.Spec) (err error) {
-	p := s.Process
-	cmdArr := c.Config.Entrypoint
-	if len(c.Config.Cmd) > 0 {
-		cmdArr = append(cmdArr, c.Config.Cmd...)
-	}
-	if c.Config.Tty != nil {
-		p.Terminal = *c.Config.Tty
-	}
-	p.Cwd = c.Config.WorkingDir
-	p.Env = c.Config.Env
-	p.Args = cmdArr
-	if c.Config.User != "" {
-		tmpArr := strings.SplitN(c.Config.User, ":", 2)
-		var u, g string
-		u = tmpArr[0]
-		if len(tmpArr) == 2 {
-			g = tmpArr[1]
-		}
-		user := &specs.User{}
-		if uid, err := strconv.Atoi(u); err == nil {
-			user.UID = uint32(uid)
-		} else {
-			user.Username = u
-		}
-		if gid, err := strconv.Atoi(g); err == nil {
-			user.GID = uint32(gid)
-		}
-	}
-
-	//TODO security config (including both seccomp and selinux)
-
-	return nil
 }
